♻️ move config to plugins
The keymaps related to Elixir filetype were removed, as they were no longer useful. In addition, the deprecated `vim.lsp.buf.signature_help` was replaced by `<leader>gh`. These changes were made to improve the keymap organization.

:recycle: Change language abbreviation from 'elixir' to 'ex' in markdown configuration

The language abbreviation in the `vim_markdown_fenced_languages` was changed to `ex` instead of `elixir` to better follow the convention.

:zap: Update lualine.nvim configuration

The Lualine configuration file was updated to center the filename and show the file status (readonly and modified status). The theme was changed to 'nightfly', and the section separators and component separators were reconfigured to better fit the theme. Finally, the 'fugitive' extension was added to display the current git branch.

:recycle: Update vim-table-mode configuration

The `table-mode-corner-corner` was updated to be '|' and the `header_fillchar` was updated to '-'. These changes were made to improve the table visualization.

:zap: Update neotest.nvim configuration

The Neotest configuration was updated to use a new `opts` table instead of running multiple functions in the 'config' option. The configuration was updated to add more options such as icons, keymaps and adapters. These new options improve the functionality and readability of the plugin.

:recycle: Update terraform.nvim configuration

The Terraform configuration was updated to enable automatic formatting on save and to align elements in a proper way. These changes were made to improve the usability of the plugin.