➕ Add hover.nvim, neotest-jest plugin
Hover.nvim is added to provide quick documentation lookup or view. This plugin binds the "K" and "gK" keys in normal mode to  a `hover preview` and `hover select`, whereas `hover preview` shows the documentation in a popup-jump list, `hover select` shows the documentation in a new buffer. It also allows the hovering and getting all types of information about code, such as LSP, GitHub, JIRA, manual pages, and dictionaries.

Also, the plugin dependencies have been modified for neotest plugin. The new dependencies is `haydenmeade/neotest-jest`.